Crate starlark_map

Ordered map optimized for starlark-rust use cases.

Modules

ordered_map

SmallMap which considers iteration order important for equality and hash.

ordered_set

SmallSet which considers order important for equality and hash.

small_map

A Map with deterministic iteration order that specializes its storage based on the number of entries to optimize memory. This is essentially IndexMap with two changes:

small_set

Ordered set.

sorted_map

SmallMap which asserts that its elements are sorted.

sorted_set

SmallSet which asserts that its elements are sorted.

sorted_vec

Vec with elements sorted.

unordered_map

HashMap that does not expose insertion order.

unordered_set

HashSet that does not expose insertion order.

vec2

A Vec<(A, B)> like object which stores A and B separately.

Macros

smallmap

Create a SmallMap from a list of key-value pairs.

smallset

Create a SmallSet from a list of values.

Structs

Hashed

A key and its hash.

StarlarkHashValue

A hash value.

StarlarkHasher

A hasher used by Starlark implementation.

StarlarkHasherBuilder

BuildHasher implementation which produces StarlarkHasher.

Traits

Equivalent

Key equivalence trait.